Add files via upload
This script captures an image of the user showing one of two symbols with their hands: open and closed fist, classifies it under the appropriate category and sends an SMS to the phone of the programmer. Twilio is needed to send an SMS - and MachineLearning4Kids was used to create the machine learning model, which was trained with many photos of open and closed fists.